requirements.txt
# Core Dependencies
langchain==1.0.8
langchain-community==0.4.1
langchain-openai==1.0.3
langchain-text-splitters==1.0.0
langchain-core==1.0.7

# Vector Database
chromadb==1.3.5

# PDF Processing
pypdf==6.3.0

# OpenAI
openai==2.8.1

# Environment Management
python-dotenv==1.2.1

# Optional: LangSmith Tracing
langsmith==0.4.45
